fixup??? grep: make PCRE2 aware of custom allocator
It is _not_ nedmalloc-specific! Never was, never will. As soon as you
use e.g. the LD_PRELOAD trick to use jemalloc, the very same issues will
arise as I reported earlier: a block that was allocated via PCRE2 wants
to be `free()`d via the custom allocator, which will cause serious
issues unless we make sure that PCRE2 uses the same allocator as Git.

Besides, the location where `pcre2_global_context` was initialized is
completely wrong. That code path was not even close to being hit in
t7816.48, however, that test case really needs that context.

Let's stop trying to be smart (and failing at it), and instead just
initialize the global context in `grep_init()` and be done with these
problems once and for all. The chase stopped being fun a month ago.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>